requirements.txt
google-search-results
requests
litellm
python-dotenv
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
pydantic>=2.0.0
supabase 